Made changes for JobScheduling sample
I changed the base image in the docker files. I was unable to build a
container that Bonsai would use, unless I specified an old enough
version of Windows Server Core.

I added some fixes to the model. At this point, I was able to train
a Bonsai brain with the model and inkling file. I was also able to wrap
the model in a container, and train with that container simulator.