Updated model after advice from Bonsai
First, I updated the NextAction parameter to be an option type, and
I changed the related Inkling code. Enumerating choices this way helps
Bonsai avoid trying to average the action, or draw conclusions about
increasing or decreasing this value. This also allows me to give Bonsai
an action mask, so I can tell Bonsai not to take certain actions.

In order to use the ActionMask, I added an extra concept to the graph.
That concept just removes the mask from the sim state, so bonsai doesn't
try to learn from those values directly.

I added a new performance measure (throughput) because I might want to
use that as the goal, instead of minimizing block time.

I also updated the model to reset the cumulative values between actions.
So throughput and block time are no longer cumulative over the episode,
just over each step.