Update goal threshold to improve training metrics
By using a non-zero threshold, the automatic training metrics become
more useful. For example, the Success metric was previously always 0%,
since it is generally impossible to have a zero-block-time schedule. Now
the Success metric moves from 0% to 100% during training.

With this change, the README no longer needs to explain why the training
metrics are so unusual, so that section is updated.